Moved header for Gosu::WindowForWrapper into the cpp file, changed if {} style, changed int function() to int function(void).
