+++ Gosu/Window.h
@@ -1,48 +1,14 @@
 #pragma once
+#include <stdbool.h>
 
 #ifdef __cplusplus
-  namespace Gosu
-  {
-    class WindowForWrapper : public Gosu::Window
-    {
-    struct Callbacks
-    {
-      void (*update)() = nullptr;
-      void (*draw)() = nullptr;
-      void (*button_down)(unsigned btn) = nullptr;
-      void (*button_up)(unsigned btn) = nullptr;
-      void (*drop)(const char* filename) = nullptr;
-      int (*needs_redraw)() = nullptr; // explicitly set to prevent segfault
-      int (*needs_cursor)() = nullptr;
-    } callbacks;
-
-    public:
-        WindowForWrapper(int width, int height, bool fullscreen, double update_interval, bool resizable);
-        void set_draw(void function());
-        void set_update(void function());
-        void set_button_down(void function(unsigned btn));
-        void set_button_up(void function(unsigned btn));
-        void set_drop(void function(const char* filename));
-        void set_needs_redraw(int function());
-        void set_needs_cursor(int function());
-
-        void update() override;
-        void draw() override;
-        void button_down(Gosu::Button btn) override;
-        void button_up(Gosu::Button btn) override;
-        void drop(const std::string& filename) override;
-        bool needs_redraw() const override;
-        bool needs_cursor() const override;
-    };
-  }
-
-  extern "C" {
+extern "C" {
 #endif
 
 typedef struct Gosu_Window Gosu_Window;
 
 // Constructor
-Gosu_Window* Gosu_Window_create(int width, int height, int fullscreen, double update_interval, int resizable);
+Gosu_Window* Gosu_Window_create(int width, int height, bool fullscreen, double update_interval, bool resizable);
 
 // callbacks
 void Gosu_Window_set_update(Gosu_Window* window, void function());
@@ -50,8 +16,8 @@ void Gosu_Window_set_draw(Gosu_Window* window, void function());
 void Gosu_Window_set_button_down(Gosu_Window* window, void function(unsigned btn));
 void Gosu_Window_set_button_up(Gosu_Window* window, void function(unsigned btn));
 void Gosu_Window_set_drop(Gosu_Window* window, void function(const char* filename));
-void Gosu_Window_set_needs_redraw(Gosu_Window* window, int function());
-void Gosu_Window_set_needs_cursor(Gosu_Window* window, int function());
+void Gosu_Window_set_needs_redraw(Gosu_Window* window, int function(void));
+void Gosu_Window_set_needs_cursor(Gosu_Window* window, int function(void));
 
 // Properties
 int Gosu_Window_width(Gosu_Window* window);
@@ -84,5 +50,5 @@ void Gosu_Window_destroy(Gosu_Window* window);
 
 
 #ifdef __cplusplus
-  }
+}
 #endif
